Add files via upload
Over the years, the company has collected basic bank details and gathered a lot of credit-related information. The management wants to build an intelligent system to segregate the people into credit score brackets to reduce manual efforts.

The objective of this project is to explore the data to identify the pattern that causes the person to become a good or bad or standard type credit score and build a machine learning model that should be able to predict or classify the credit score type.